+++ test/cards/test_cards.py
@@ -1,5 +1,4 @@
 from typing import (
-    Optional,
     Sequence,
 )
 from unittest.mock import Mock, patch
@@ -24,13 +23,13 @@ from balatro_gym.cards.planet import Mercury, Pluto
 from balatro_gym.cards.voucher import ClearanceSale, Liquidation, Voucher
 from balatro_gym.game.scoring import get_poker_hand, score_hand
 from balatro_gym.interfaces import BoardState, PokerHand, PokerHandType
-from balatro_gym.testing_utils import make_card
+from test.utils import _make_card
 
 
 @pytest.mark.unit
 def test_enhancement_bonus() -> None:
     enhancement = BonusCard()
-    card = make_card(enhancement=enhancement)
+    card = _make_card(enhancement=enhancement)
     assert card.enhancement == enhancement
     assert card.get_chips() == 50 + card._base_chips
 
@@ -38,7 +37,7 @@ def test_enhancement_bonus() -> None:
 @pytest.mark.unit
 def test_enhancement_mult() -> None:
     enhancement = MultCard()
-    card = make_card(enhancement=enhancement)
+    card = _make_card(enhancement=enhancement)
     assert card.enhancement == enhancement
     assert card.get_mult() == 4
 
@@ -46,7 +45,7 @@ def test_enhancement_mult() -> None:
 @pytest.mark.unit
 def test_enhancement_wild() -> None:
     enhancement = WildCard()
-    card = make_card(enhancement=enhancement)
+    card = _make_card(enhancement=enhancement)
     assert card.enhancement == enhancement
     assert len(set(card.suit)) == len(Suit)
 
@@ -54,7 +53,7 @@ def test_enhancement_wild() -> None:
 @pytest.mark.unit
 def test_enhancement_glass() -> None:
     enhancement = GlassCard()
-    card = make_card(enhancement=enhancement)
+    card = _make_card(enhancement=enhancement)
     assert card.enhancement == enhancement
     assert card.get_multiplication() == 2
 
@@ -62,7 +61,7 @@ def test_enhancement_glass() -> None:
 @pytest.mark.unit
 def test_enhancement_glass_scoring() -> None:
     enhancement = GlassCard()
-    card = make_card(enhancement=enhancement)
+    card = _make_card(enhancement=enhancement)
     board = BoardState()
     board.jokers = []
     blind_mock = Mock()
@@ -78,7 +77,7 @@ def test_enhancement_glass_scoring() -> None:
 @pytest.mark.parametrize("probability_modifier", [1, 2, 10])
 def test_enhancement_glass_destroy(probability_modifier: int) -> None:
     enhancement = GlassCard()
-    card = make_card(enhancement=enhancement)
+    card = _make_card(enhancement=enhancement)
     deck = Deck([card])
     submitted_hand = deck.deal(1)
     assert card in deck.cards_played
@@ -99,7 +98,7 @@ def test_enhancement_glass_destroy(probability_modifier: int) -> None:
 @pytest.mark.unit
 def test_enhancement_steel() -> None:
     enhancement = SteelCard()
-    card = make_card(enhancement=enhancement)
+    card = _make_card(enhancement=enhancement)
     assert card.enhancement == enhancement
     assert card.get_multiplication() == 1.5
 
@@ -108,8 +107,8 @@ def test_enhancement_steel() -> None:
 def test_enhancement_steel_scoring() -> None:
     # Submit a single card as a hand with a single steel card held in hand
     enhancement = SteelCard()
-    held_card = make_card(enhancement=enhancement)
-    submitted_card = make_card()
+    held_card = _make_card(enhancement=enhancement)
+    submitted_card = _make_card()
     submitted_hand = [submitted_card]
     board = BoardState()
     board.jokers = []
@@ -129,7 +128,7 @@ def test_enhancement_steel_scoring() -> None:
 @pytest.mark.unit
 def test_enhancement_stone() -> None:
     enhancement = StoneCard()
-    card = make_card(enhancement=enhancement)
+    card = _make_card(enhancement=enhancement)
     assert card.enhancement == enhancement
     assert enhancement.get_chips() == 50
     assert card.get_chips() == 50
@@ -140,7 +139,7 @@ def test_enhancement_stone() -> None:
 def test_enhancement_stone_scoring() -> None:
     # Submit a single stone card
     enhancement = StoneCard()
-    card = make_card(enhancement=enhancement)
+    card = _make_card(enhancement=enhancement)
     board = BoardState()
     board.jokers = []
     blind_mock = Mock()
@@ -155,7 +154,7 @@ def test_enhancement_stone_scoring() -> None:
 @pytest.mark.unit
 def test_enhancement_gold() -> None:
     enhancement = GoldCard()
-    card = make_card(enhancement=enhancement)
+    card = _make_card(enhancement=enhancement)
     assert card.enhancement == enhancement
     assert card.get_scored_money() == 0
     assert card.get_end_money() == 3
@@ -164,7 +163,7 @@ def test_enhancement_gold() -> None:
 @pytest.mark.unit
 def test_enhancement_lucky() -> None:
     enhancement = LuckyCard()
-    card = make_card(enhancement=enhancement)
+    card = _make_card(enhancement=enhancement)
     assert card.enhancement == enhancement
     with patch.object(balatro_gym.cards.interfaces, "np") as mock:
         random_mock = Mock()
@@ -182,7 +181,7 @@ def test_enhancement_lucky() -> None:
 @pytest.mark.parametrize("probability_modifier", [1, 2, 10])
 def test_enhancement_lucky_modifiers(probability_modifier: int) -> None:
     enhancement = LuckyCard()
-    card = make_card(enhancement=enhancement)
+    card = _make_card(enhancement=enhancement)
     assert card.enhancement == enhancement
     with patch.object(balatro_gym.cards.interfaces, "np") as mock:
         random_mock = Mock()
