+++ test/utils.py
@@ -1,0 +1,13 @@
+from typing import Optional
+
+from balatro_gym.cards.interfaces import Edition, Enhancement, PlayingCard, Rank, Seal, Suit
+
+
+def _make_card(
+    rank: Rank = Rank.ACE,
+    suit: Suit = Suit.HEARTS,
+    enhancement: Optional[Enhancement] = None,
+    edition: Optional[Edition] = None,
+    seal: Optional[Seal] = None,
+) -> PlayingCard:
+    return PlayingCard(rank, suit, enhancement, edition, seal)
